+++ src/main/java/com/axonactive/basketball/api/TeamResources.java
@@ -1,16 +1,22 @@
 package com.axonactive.basketball.api;
 
+import com.axonactive.basketball.entities.Arena;
 import com.axonactive.basketball.entities.Team;
+import com.axonactive.basketball.enums.Conference;
 import com.axonactive.basketball.exceptions.ResourceNotFoundException;
 import com.axonactive.basketball.services.dtos.TeamDTO;
+import com.axonactive.basketball.services.impl.ArenaServiceImpl;
 import com.axonactive.basketball.services.impl.TeamServiceImpl;
 import com.axonactive.basketball.services.mappers.TeamMapper;
+import com.axonactive.basketball.services.requests.TeamRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping(TeamResources.PATH)
@@ -18,35 +24,63 @@ public class TeamResources {
     public static final String PATH = "/api/team";
     @Autowired
     TeamServiceImpl teamService;
+    @Autowired
+    ArenaServiceImpl arenaService;
+
     @GetMapping
-    public ResponseEntity<List<TeamDTO>> findAll(){
+    public ResponseEntity<List<TeamDTO>> findAll() {
         return ResponseEntity.ok(TeamMapper.INSTANCE.toDTOs(teamService.findAll()));
     }
+
     @GetMapping("/{name}")
-    public ResponseEntity<TeamDTO> findByID(@PathVariable(value = "name") String name) throws ResourceNotFoundException{
-        Team team = teamService.findByID(name)
-                .orElseThrow(() -> new ResourceNotFoundException("Name not found: " + name));
-        return ResponseEntity.ok(TeamMapper.INSTANCE.toDTO(team));
+    public ResponseEntity<?> findByID(@PathVariable(value = "name") String name) {
+        Optional<Team> team = teamService.findByID(name);
+        if (team.isPresent())
+            return ResponseEntity.ok(TeamMapper.INSTANCE.toDTO(team.get()));
+        else return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Team name not found: " + name);
     }
+
     @PostMapping
-    public ResponseEntity<TeamDTO> create(@RequestBody Team team){
-        return ResponseEntity.created(URI.create(PATH + "/" + team.getName())).body(TeamMapper.INSTANCE.toDTO(teamService.save(team)));
+    public ResponseEntity<?> create(@RequestBody TeamRequest teamRequest) {
+        Optional<Arena> arena = arenaService.findByID(teamRequest.getArenaName());
+        if (!arena.isPresent())
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Arena name not found: " + teamRequest.getArenaName());
+        else {
+            Team team = new Team(teamRequest.getName(),
+                    teamRequest.getLocation(),
+                    teamRequest.getDateFound(),
+                    teamRequest.getSalaryCap(),
+                    Conference.valueOf(teamRequest.getConference()),
+                    arena.get()
+            );
+            return ResponseEntity.created(URI.create(PATH + "/" + team.getName())).body(TeamMapper.INSTANCE.toDTO(teamService.save(team)));
+        }
     }
+
     @PutMapping("/{name}")
-    public ResponseEntity<TeamDTO> update(@PathVariable(value = "name") String name,
-                                          @RequestBody Team teamDetails) throws ResourceNotFoundException{
-        Team team = teamService.findByID(name).orElseThrow(() -> new ResourceNotFoundException("Name not found: " + name));
-        team.setArena(teamDetails.getArena());
-        team.setLocation(teamDetails.getLocation());
-        team.setDateFound(teamDetails.getDateFound());
-        team.setSalaryCap(teamDetails.getSalaryCap());
-        team.setConference(teamDetails.getConference());
-        return ResponseEntity.ok(TeamMapper.INSTANCE.toDTO(teamService.save(team)));
+    public ResponseEntity<?> update(@PathVariable(value = "name") String name,
+                                    @RequestBody TeamRequest teamRequest) throws ResourceNotFoundException {
+        Optional<Team> team = teamService.findByID(name);
+        Optional<Arena> arena = arenaService.findByID(teamRequest.getArenaName());
+        if (!arena.isPresent())
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Arena name not found: " + teamRequest.getArenaName());
+        if (team.isPresent()) {
+            team.get().setArena(arena.get());
+            team.get().setLocation(teamRequest.getLocation());
+            team.get().setDateFound(teamRequest.getDateFound());
+            team.get().setSalaryCap(teamRequest.getSalaryCap());
+            team.get().setConference(Conference.valueOf(teamRequest.getConference()));
+            return ResponseEntity.ok(TeamMapper.INSTANCE.toDTO(teamService.save(team.get())));
+        } else return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Team name not found: " + name);
     }
+
     @DeleteMapping("/{name}")
-    public ResponseEntity<Void> deleteByID(@PathVariable(value = "name") String name) throws ResourceNotFoundException{
-        Team team = teamService.findByID(name).orElseThrow(() -> new ResourceNotFoundException("Name not found: " + name));
-        teamService.deleteByID(name);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<?> deleteByID(@PathVariable(value = "name") String name) throws ResourceNotFoundException {
+        Optional<Team> team = teamService.findByID(name);
+        if (team.isPresent()) {
+            teamService.deleteByID(name);
+            return ResponseEntity.ok("Successfully deleted");
+        }
+        else return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Team name not found: " + name);
     }
 }
